requirements.txt
# your virtual env folder:
flask_venv/
# the instance/ folder containing our secret stuff!
instance/
# anything else you want git to ignore, add here:# your virtual env folder:
flask_venv/
# the instance/ folder containing our secret stuff!
instance/
# anything else you want git to ignore, add here:# your virtual env folder:
flask_venv/
# the instance/ folder containing our secret stuff!
instance/
# anything else you want git to ignore, add here:# your virtual env folder:# your virtual env folder:# your virtual env folder:# your virtual env folder:# your virtual env folder:# your virtual env folder:# your virtual env folder:# your virtual env folder:# your virtual env folder:# your virtual env folder:# your virtual env folder:# your virtual env folder:# your virtual env folder:# your virtual env folder:# your virtual env folder:# your virtual env folder:# your virtual env folder:# your virtual env folder:# your virtual env folder:# your virtual env folder:# your virtual env folder:# your virtual env folder:# your virtual env folder:# your virtual env folder:# your virtual env folder:
flask_venv/
# the instance/ folder containing our secret stuff!
instance/
# anything else you want git to ignore, add here:# your virtual env folder:
flask_venv/
# the instance/ folder containing our secret stuff!
instance/
# anything else you want git to ignore, add here:pip
Flask
cassandra-driver